+++ Project-tokens-Algoritmos-de-ordenamiento/Project-tokens-Algoritmos-de-ordenamiento/Fichas.cbp
@@ -31,6 +31,18 @@
 		<Compiler>
 			<Add option="-Wall" />
 		</Compiler>
+		<Unit filename="AlgoritmoBusqueda.c">
+			<Option compilerVar="CC" />
+		</Unit>
+		<Unit filename="AlgoritmoBusqueda.h" />
+		<Unit filename="ManejoArchivos.c">
+			<Option compilerVar="CC" />
+		</Unit>
+		<Unit filename="ManejoArchivos.h" />
+		<Unit filename="Tablero.c">
+			<Option compilerVar="CC" />
+		</Unit>
+		<Unit filename="Tablero.h" />
 		<Unit filename="main.c">
 			<Option compilerVar="CC" />
 		</Unit>
