+++ Project-tokens-Algoritmos-de-ordenamiento/Project-tokens-Algoritmos-de-ordenamiento/ManejoArchivos.h
@@ -1,0 +1,7 @@
+#ifndef MANEJOARCHIVOS_H
+#define MANEJOARCHIVOS_H
+void abrirArchivo();
+int revisaArchivo(FILE* archivo);
+int revisaArray(int N, int arreglo[]);
+
+#endif // MANEJOARCHIVOS_H
