+++ Project-tokens-Algoritmos-de-ordenamiento/Project-tokens-Algoritmos-de-ordenamiento/PruebasGrupo1.txt
@@ -1,5 +1,5 @@
 5
-1 2 2 4 3 4 5 1 5 3
+5 1 1 2 2 3 3 4 4 4
 2
 1 1 1 2
 3
