+++ Project-tokens-Algoritmos-de-ordenamiento/Project-tokens-Algoritmos-de-ordenamiento/bb.h
@@ -1,0 +1,16 @@
+#ifndef BB_H
+#define BB_H
+
+
+class bb
+{
+    public:
+        bb() {}
+        virtual ~bb() {}
+
+    protected:
+
+    private:
+};
+
+#endif // BB_H
